Extract Gateway WebSocket lifecycle management
